pkg/resolvers: fix message length mismatch
In previous commit, we change from io.ReadAll to io.ReadFull, to keep
the io.EOF error. However, the message length sent by client may not
have the same length as the response message from server. The quic-go
example can use io.ReadFull, because it always knows exactly the length
of the response message.

Instead, we can detect if we got an empty buf in case of no error, and
return io.EOF error to the caller instead.